[keyserver] Prevent creation of sidebars from pinned message robotext
Summary:
These sidebars are causing our keyserver to completely crash and the app to break ([ENG-4849](https://linear.app/comm/issue/ENG-4849/thread-from-robotext-about-pinned-message-causing-errors)). This diff prevents the creation of such sidebars on the keyserver. Preventing the sidebar creation button from appearing in the UI on `native` and `web` will be handled separately.

Depends on D9231

Test Plan: Try creating a sidebar from a pinned message in my local environment and confirm that it now errors out

Reviewers: atul, rohan, inka

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D9232