[lib] introduce createPendingPersonalThread function
Summary:
This diff introduces a helper function called `createPendingPersonalThread`. At the moment we are going to use this function if we search for a user that we don't have a preexisting shared personal thread yet, but we will also need to do the same thing for user profiles (if we view a user profile that we don't have a preexisting shared personal thread yet).

Depends on D9120

Test Plan: `flow` and confirmed that there are no regressions with the search experience when I try and search for a user that I don't already have a thread with yet

Reviewers: atul, inka, ashoat, tomek

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9235