[keyserver] Skip early exit in deleteThread if ignorePermissions
Summary:
When I ran the migration in D9230 on production, it succeeded but didn't delete the threads. After some testing I realized there was an error in D9229, which didn't come up in my testing because I was testing in a community that `commbot` happened to be part of in my local environment.

The issue is that the "early exit" here will trigger if there is no permissions blob. In the case of `ignorePermissions`, we want the operation to go through even if there is no permissions blob for the script viewer's user (`commbot`).

Test Plan: I applied this patch in production, and then re-ran the migration. The bad threads were finally deleted

Reviewers: atul, rohan, inka

Reviewed By: rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D9234