[lib] Extract the unsupported sidebar source message types to one single source of truth
Summary:
As mentioned in the previous diff, we want a single source of truth for a list of message types that should disallow sidebar creation. This diff moves the check into a helper function in
`message-utils.js`, and adds a quick unit test alongside this to check the integrity of the method.

Depends on D9237

Addresses part of [[ https://linear.app/comm/issue/ENG-4849/thread-from-robotext-about-pinned-message-causing-errors#comment-96a285ec | ENG-4849 ]]

Test Plan: Ran `flow`, confirmed that the test plan in D9237 still works, and ran `yarn workspace lib test` to see the unit tests passing

Reviewers: ashoat, atul, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D9238