[native] Use threadInfo in setClientDBStoreActionType
Summary:
Part of [ENG-4959](https://linear.app/comm/issue/ENG-4959)

The hashes will be kept in the `threadStore` alongside the `threadInfos`, but will be persisted in the redux persist. Because of that we need to change the action to only contain `threadInfos` and not the whole `threadStore`.

Explanation why the hashes should be stored here:
- we need to store the hashes because generating hashes for all threads would take ~20s, so generating them at each app startup is not feasible
- we could store them in `RawThreadInfo` but that would mean much larger changes in the codebase. It also adds some complexity (we generate hash from `RawThreadInfo` but the hash is contained inside of it?)
- we could store it in some other `hashStore`, but that would mean we need larger changes in `thread-store-ops.js` (where we will keep the hashes up to date with the current `threadInfos`)

Test Plan: Open the app as a logged in user, check that the threads are still there. In redux devtools check that the action was sent correctly

Reviewers: kamil, marcin, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9255