[lib] Move entries updates logic into a spec
Summary: Move the logic into specs. Entry store update is based on mutating the input, so it isn't ideal, but I didn't want to alter the logic significantly as a part of this simple refactoring.

Depends on D9245

https://linear.app/comm/issue/ENG-4241/handle-processupdatesactiontype-as-a-part-of-a-spec

Test Plan: Check if joining a thread and updating an entry correctly affect the store.

Reviewers: kamil, bartek, inka

Subscribers: michal