[lib/web/native] Add IntegrityStore
Summary:
Part of [ENG-4959](https://linear.app/comm/issue/ENG-4959)
Suggested in D9255, which was a previous attempy at storing hashes in redux

We want to store thread hashes in redux so that we don't have to calculate all of them whenever the state check is triggered. Instead we keep them in sync using thread ops. Additionaly the hashes will be now generated based on the client schema so we convert them on the server instead of the client.

Test Plan:
- Migrate native web app -> check if `integrityStore` contains hashes
- Reload native web app -> check if hashes are still there
- Open web app -> check if initial redux state contained new hashes and that they were added to redux
- On both web and native apps -> change thread color and check if the thread hash was updated

Log in performance:
- iOS emulator takes about 1ms per one thread and the speed should be O(n). In the case from the linear task where the state check took 53s, there was about 2000 threads, so logging in would take additional 2s
	- Even if we lazily generate the hashes, I don't think there would be a lot of thread updated between logging in and first state check so all that would do is moving the wait time from logging in to the first state check (with a frozen app), so I'm not sure if it's worth it
- Web is even faster, ~2000 threads should take about 0.3s

Reviewers: tomek, atul, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9265