[native] introduce useNavigateToUserProfileBottomSheet hook
Summary:
This diff introduces the `useNavigateToUserProfileBottomSheet` hook which is going to be used to navigate to the user profile bottom sheet screen. This logic will need to be repeated in every place we want the user profiles to be accessible, so I thought that instead of copy+pasting the logic we should preemptively factor out this logic into this helper hook.

Depends on D9383

Test Plan:
Confirmed that I can navigate to the user profiles screen using this hook

{F769182}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9261