[services-lib] Introduce AuthService for token verification
Summary:

Created a helper "service struct" to easily handle token operations. It allows to obtain a service-to-service token and to verify a token's validity.
For `UserIdentity` (client tokens) it calls Identity Service to verify the token.

This struct is designed to be provided as app-data to HTTP actix servers - it's going to be used by auth middleware.

Depends on D9279

Test Plan:

cargo check, logic is tested together with next diffs

Reviewers: varun, michal

Subscribers: