[lib] Get rid of innerNotificationTexts
Summary:
When @tomek first started using message specs for notifs in D644, he introduced this parameter because it was needed for sidebar sources.

However, at a later point we realized that we had notifs disabled for sidebar sources. In D6870 I replaced that code with an `invariant`.

We no longer need this `innerNotificationTexts`, and it presents an ongoing maintenance cost (most recently noticed in D9249). As such, I'd like to remove it. We can always bring it back later if we need it.

Test Plan: Flow

Reviewers: tomek, rohan

Subscribers: