[lib] Only consider children of GENESIS when proposing existing chats
Summary:
This diff addresses [ENG-5016](https://linear.app/comm/issue/ENG-5016/commbot-chat-appears-when-trying-to-message-atul-and-varun).

The chat composer on both `native` and `web` allows the viewer to select a set of users they want to chat with. As new users are added to the list, the chat composer attempts to find an existing chat that includes that exact set of users. Note that if a match is made, the viewer is unable to create a new chat; the only option is to use the existing chat.

The problem is that the current code is returning channels inside specific communities. We consider it okay to have two channels in a community with the same membership since they might have different topics. On top of that, matching a channel within a specific community is confusing behavior when the chat composer is not meant to be community-specific.

This diff addresses the issue by making sure we only propose children of `GENESIS`. More context in the commentary on Linear.

Test Plan: I patched this onto my keyserver in production and confirmed that I was no longer seeing the commbot chat appear when I wanted to chat with Atul and Varun

Reviewers: atul, ginsu, tomek

Subscribers: