Introduce Brotli compression for keyserver -> client socket messages
Summary:
This addresses [ENG-5018](https://linear.app/comm/issue/ENG-5018/compress-large-keyserver-client-payloads), which is one of the follow-ups to [ENG-5007](https://linear.app/comm/issue/ENG-5007/client-socket-cant-connect-to-keyserver).

Scoping this just to staff members for now. If this goes well, we can try the same thing for client -> keyserver socket messages and for HTTP REST endpoints as well.

Depends on D9293

Test Plan: I tested sending a UTF-8 message (emoji with skin tone) from web to native and from native to web. I then forced a socket reconnection for the sender, and confirmed that the message didn't change once the socket was established.

Reviewers: atul, michal, kamil

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D9294