[lib] Move calendar filters update reducer into specs
Summary: Changed a solution slightly. Instead of mutating a set for each update, we're creating a new one, which might be less performant, but is better for maintainability.

Depends on D9296

https://linear.app/comm/issue/ENG-4241/handle-processupdatesactiontype-as-a-part-of-a-spec

Test Plan: Bring back the old code and check if it produces the same result while handling delete, join, or update thread update.

Reviewers: kamil, bartek, inka

Subscribers: michal