[keyserver] Pass `--performance-schema` to MariaDB
Summary:
As I mentioned in D9304 it turns out we can totally just pass `--performance-schema` to `mariadb`. I was misled by some old MySQL (not MariaDB) docs as well as a comment someone left on GitHub that inaccurately stated that `performance-schema` had to be defined in a `.cnf` file and couldn't be passed as command line argument when started up. Annoying, but this is a much cleaner solution anyways.

---

Depends on D9304

Test Plan:
Confirmed that `performance_schema` was set as expected (`docker-compose up`):

Reviewers:
ashoat, tomek, ginsu, rohan