[lib] Update types for SIDEBAR_SOURCE messages to reflect isInvalidSidebarSource
Summary:
Now that `isInvalidSidebarSource` is currently the single source of truth for the valid/invalid message types that can be sidebar sources, it makes sense to update the types for
`SidebarSourceMessageInfo` `SidebarSourceMessageData` and `RawSidebarSourceMessageInfo` to reflect this.

Resolves https://linear.app/comm/issue/ENG-5207/update-types-for-sidebar-source-messages-to-reflect

Depends on D9322

Test Plan: Ran `flow` to confirm that updating the `sourceMessage` types didn't cause issues, and manually tested sidebar creation again for test / change role messages / toggle pin messages.

Reviewers: atul, ginsu, ashoat

Reviewed By: ginsu, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9324