[lib] Introduce a canBeSidebarSource function in each message spec
Summary:
The last part of unifying the logic and source of truth for determining eligible messages to be sidebar sources is to have a `canBeSidebarSource` function in each message spec. This diff introduces
the method to each existing message spec, and the next diffs will replace individual calls to `isInvalidSidebarSoruce` to `messageSpecs[rawMessageInfo.type].canBeSidebarSource(...)`. This ensures that we'll
only ever need to update the function in one place if we ever want to change this.

Depends on D9324

Addresses 1c of [[https://linear.app/comm/issue/ENG-4849/thread-from-robotext-about-pinned-message-causing-errors#comment-36f2e045 | ENG-4849]].

Test Plan: Ran `flow` and manually confirmed the output of `canBeSidebarSoruce` matches the expected output of the unit tests

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9325