[keyserver] Test impact of skipping old sidebars in Ashoat's ThreadStore
Summary: This diff stack introduces a script that tests various different approaches to reducing the size of the `ThreadStore`. We may or may not land it... the goal is just to get some numbers.

Depends on D9335

Test Plan: Run the script:
```
after filtering out old sidebars, Ashoat's ThreadStore is 28832851 bytes large
```

Corresponds to 27.50 MiB, which is a reduction of 60.16%.

Reviewers: atul, tomek

Subscribers: