[keyserver] Test impact of improving encoding on Ashoat's ThreadStore
Summary:
This diff stack introduces a script that tests various different approaches to reducing the size of the `ThreadStore`. We may or may not land it... the goal is just to get some numbers.

Depends on D9336

Test Plan:
Run the script:
```
after improving encoding, Ashoat's ThreadStore is 11123335 bytes large
```

Corresponds to 10.61 MiB, which is a reduction of 84.63%.

Reviewers: atul, tomek

Differential Revision: https://phab.comm.dev/D9337