[native] new method to initialize content olm session
Summary:
We need this method to intialize a content session and send the initial encrypted message to the keyserver

Depends on D9326

Test Plan: called the new method with the keys from my local keyserver. was able to create an olm session by sending these via the new auth responder later in the stack.

Reviewers: kamil, bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9334