[keyserver] Test impact of introduction of isAdminRole on Ashoat's ThreadStore
Summary:
This diff stack introduces a script that tests various different approaches to reducing the size of the `ThreadStore`. We may or may not land it... the goal is just to get some numbers.

Depends on D9337

Test Plan:
Run the script:
```
after introducing isAdminRole, Ashoat's ThreadStore is 14839491 bytes large
```

Corresponds to 14.15 MiB, which is a reduction of 79.50%.

Reviewers: atul, tomek

Differential Revision: https://phab.comm.dev/D9338