[keyserver] Test impact of introducing ThreadRolePermissionsStore on Ashoat's ThreadStore
Summary:
This diff stack introduces a script that tests various different approaches to reducing the size of the `ThreadStore`. We may or may not land it... the goal is just to get some numbers.

Depends on D9338

Test Plan:
Run the script:
```
after introducing ThreadRolePermissionsStore, Ashoat's ThreadStore is 5640441 bytes large
```

Corresponds to 5.38 MiB, which is a reduction of 92.21%.

Reviewers: atul, tomek

Reviewed By: atul

Differential Revision: https://phab.comm.dev/D9339