[native] introduce UserProfileAvatarModal
Summary:
This diff introduces the User Profile Avatar Modal component. The User Profile Avatar Modal uses the new FullScreenViewModal and XXL avatar size that I introduced earlier.

This is step 4 in the list below
Outlined below are the steps I will take in this stack (each point here will be it's own diff):
1. Replace out of box `BottomSheetModal` with basic `BottomSheet` (Unfortunately, BottomSheetModal does not play nice with nested navigator)
2. Polish up the BottomSheetBackdrop animation
3. Introduce User Profile Bottom Sheet Navigators (this is so we can navigate to User Profile Avatar Modal from the User Profile Bottom Sheet)
4. Introduce the User Profile Avatar Modal component and all the necessary things (like route names) that we will need to navigate to this new screen
5. Factor out the User Avatar component in `UserProfile` into it's own separate component to keep things better organized

Depends on D9375

Test Plan:
Confirmed that I was able to navigate to the User Profile Avatar Modal screen (the next diff will introduce the navigation method to this screen), and I can do things like zoom, pan, etc with the user avatars

{F791932}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9376