[native] update avatar size variants to use t-shirt sizes
Summary: For the user profile avatar view, I plan to introduce a new avatar size. This was something @atul called out in D8913. Figured before I introduce this new size, it would be worthwhile now to address this feedback

Test Plan: `flow` and confirmed that there were no regressions with the avatar sizes

Subscribers: ashoat, tomek, wyilio, atul

Differential Revision: https://phab.comm.dev/D9364