[lib/native] lift AvatarSize type into lib
Summary:
This `AvatarSize` type is only accessible in the `native` codebase. However, we will need this for `web` as well so it makes sense to lift this type into `lib`. This diff just handles the lifting from `native` to `lib`, a subsequent diff will handle the actual usage of this type in `web`

Depends on D9364

Test Plan: `flow`

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9365