[native] introduce children prop to FullScreenViewModal
Summary: To view a full screen avatar for user profiles we are going to want to reuse parts of the `ImageModal` since it has all the same zooming/panning/gesture functionality we need. However, right now there are parts of `ImageModal` that are specic to viewing a mulitmeida message. We need to factor those parts out, so I created `FullScreenViewModal`

Test Plan: Confirmed that there were no regressions with the viewing an image multimedia message experience

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9359