[Tunnelbroker] handle all types of socket messages
Summary:
This is an enhancement, we should handle those messages properly.

Ping/Pong should be handled automatically, eg. You can send a `ping` message to WebSocket created with `ws` which we use on Keyserver, and `pong` will be sent implicitly, even if it's not implemented in the top level.

Based on the tool we now plan to use this is not super needed but it's better to avoid closing a socket when the client library/tool sends a `ping`.

Test Plan:
Run tests, and verify those tests test it by checking Tunnelbroker logs.

Depends on D9460

Reviewers: michal, bartek, varun, jon

Reviewed By: michal, bartek

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9461