[nix] Create new url facts
Summary:
Part of [ENG-5153](https://linear.app/comm/issue/ENG-5153)
Depends on D9397

We would like to separate hosting web app on dev environment to `localhost:3000/webapp/`` and move the keyserver logic to `localhost:3000/keyserver/` (so in the end nothing will be hosted on `localhost:3000/`) to simplify the dev environment and allow for web <-> keyserver separation. We can also use this opportunity to cleanup the naming.  This diff automatically adds new files on nix start. They will be used in the later diff.

Test Plan:
Run `nix develop`, check if files are generated

Reviewers: inka, kamil, atul

Subscribers: