[web/native/keyserver] Start using new url facts
Summary:
Part of [ENG-5153](https://linear.app/comm/issue/ENG-5153)
Depends on D9398

Now we can start using the new `/webapp/` and `/keyserver` urls in other places.
This also brings back the option of testing the web app with "staging" keyserver because it sets the `urlPrefix` to the one specified in the url facts (passed as a variable in html) (more context in D9145).

Test Plan:
- check that web and native migrations work
- check that uploading a new image works
- test that navigating to `https://localhost:3000/comm/chat/thread/256%7C83932/` redirects to `/webapp/`
- I run the keyserver docs according to the instructions in the `nix_keyserver_deployment.md` to check if both new web app and keyserver endpoints work:
 - The webapp successfully loaded under `http://localhost:3000/webapp/`
 - In the network inspector I could see a successful request to the keyserver (`http://localhost:3000/keyserver/get_initial_redux_state`)

Reviewers: inka, kamil, atul, ashoat

Reviewed By: kamil, ashoat

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9451