[web] Make worker and wasm urls relative
Summary:
Part of [ENG-5153](https://linear.app/comm/issue/ENG-5153)

We would like to separate hosting web app on dev environment to `localhost:3000/webapp/` and the keyserver logic to `localhost:3000/keyserver/`. This means that we want to start we need to replace absolute URL (that point to `localhost:3000/`) with relative ones.

Depends on D9396

Test Plan: Run `yarn dev` and `yarn prod` and check that both notif and database workers are downloaded with `/comm/workers/` (this will be changed in the later diffs). Also make sure that on `yarn prod` opaque-ke and olm wasm files are also loaded through `/comm/`

Reviewers: inka, kamil, atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9397